refactor: extract shared mqtt packet dispatcher flow
