+++ mqtt-cs/src/main/java/org/apache/rocketmq/mqtt/cs/protocol/mqtt/MqttPacketDispatcher.java
@@ -20,21 +20,18 @@ package org.apache.rocketmq.mqtt.cs.protocol.mqtt;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.handler.codec.mqtt.MqttConnectMessage;
 import io.netty.handler.codec.mqtt.MqttMessage;
 import io.netty.handler.codec.mqtt.MqttPubAckMessage;
 import io.netty.handler.codec.mqtt.MqttPublishMessage;
 import io.netty.handler.codec.mqtt.MqttSubscribeMessage;
 import io.netty.handler.codec.mqtt.MqttUnsubscribeMessage;
-import io.netty.util.ReferenceCountUtil;
 import org.apache.rocketmq.mqtt.common.hook.HookResult;
 import org.apache.rocketmq.mqtt.common.hook.UpstreamHookManager;
 import org.apache.rocketmq.mqtt.common.model.MqttMessageUpContext;
 import org.apache.rocketmq.mqtt.common.util.HostInfo;
-import org.apache.rocketmq.mqtt.cs.channel.ChannelDecodeException;
-import org.apache.rocketmq.mqtt.cs.channel.ChannelException;
 import org.apache.rocketmq.mqtt.cs.channel.ChannelInfo;
+import org.apache.rocketmq.mqtt.cs.protocol.AbstractMqttPacketDispatcher;
 import org.apache.rocketmq.mqtt.cs.protocol.mqtt.handler.MqttConnectHandler;
 import org.apache.rocketmq.mqtt.cs.protocol.mqtt.handler.MqttDisconnectHandler;
 import org.apache.rocketmq.mqtt.cs.protocol.mqtt.handler.MqttPingHandler;
@@ -50,12 +47,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
-import java.util.concurrent.CompletableFuture;
 
 
 @ChannelHandler.Sharable
 @Component
-public class MqttPacketDispatcher extends SimpleChannelInboundHandler<MqttMessage> {
+public class MqttPacketDispatcher extends AbstractMqttPacketDispatcher {
     private static Logger logger = LoggerFactory.getLogger(MqttPacketDispatcher.class);
 
     @Resource
@@ -92,58 +88,17 @@ public class MqttPacketDispatcher extends SimpleChannelInboundHandler<MqttMessag
     private UpstreamHookManager upstreamHookManager;
 
     @Override
-    protected void channelRead0(ChannelHandlerContext ctx, MqttMessage msg) throws Exception {
-        if (!ctx.channel().isActive()) {
-            return;
-        }
-        if (!msg.decoderResult().isSuccess()) {
-            throw new ChannelDecodeException(ChannelInfo.getClientId(ctx.channel()) + "," + msg.decoderResult());
-        }
-        ChannelInfo.touch(ctx.channel());
-        boolean preResult = preHandler(ctx, msg);
-        if (!preResult) {
-            return;
-        }
-        CompletableFuture<HookResult> upstreamHookResult;
-        try {
-            if (msg instanceof MqttPublishMessage) {
-                ((MqttPublishMessage) msg).retain();
-            }
-            upstreamHookResult = upstreamHookManager.doUpstreamHook(buildMqttMessageUpContext(ctx), msg);
-            if (upstreamHookResult == null) {
-                _channelRead0(ctx, msg, null);
-                return;
-            }
-        } catch (Throwable t) {
-            logger.error("", t);
-            if (msg instanceof MqttPublishMessage) {
-                ReferenceCountUtil.release(msg);
-            }
-            throw new ChannelException(t.getMessage());
-        }
-        upstreamHookResult.whenComplete((hookResult, throwable) -> {
-            if (msg instanceof MqttPublishMessage) {
-                ReferenceCountUtil.release(msg);
-            }
-            if (throwable != null) {
-                logger.error("", throwable);
-                ctx.fireExceptionCaught(new ChannelException(throwable.getMessage()));
-                return;
-            }
-            if (hookResult == null) {
-                ctx.fireExceptionCaught(new ChannelException("UpstreamHook Result Unknown"));
-                return;
-            }
-            try {
-                _channelRead0(ctx, msg, hookResult);
-            } catch (Throwable t) {
-                logger.error("", t);
-                ctx.fireExceptionCaught(new ChannelException(t.getMessage()));
-            }
-        });
+    protected Logger logger() {
+        return logger;
     }
 
-    private void _channelRead0(ChannelHandlerContext ctx, MqttMessage msg, HookResult upstreamHookResult) {
+    @Override
+    protected UpstreamHookManager upstreamHookManager() {
+        return upstreamHookManager;
+    }
+
+    @Override
+    protected void doChannelRead(ChannelHandlerContext ctx, MqttMessage msg, HookResult upstreamHookResult) {
         switch (msg.fixedHeader().messageType()) {
             case CONNECT:
                 mqttConnectHandler.doHandler(ctx, (MqttConnectMessage) msg, upstreamHookResult);
@@ -179,7 +134,8 @@ public class MqttPacketDispatcher extends SimpleChannelInboundHandler<MqttMessag
         }
     }
 
-    private boolean preHandler(ChannelHandlerContext ctx, MqttMessage msg) {
+    @Override
+    protected boolean preHandler(ChannelHandlerContext ctx, MqttMessage msg) {
         switch (msg.fixedHeader().messageType()) {
             case CONNECT:
                 return mqttConnectHandler.preHandler(ctx, (MqttConnectMessage) msg);
@@ -206,7 +162,8 @@ public class MqttPacketDispatcher extends SimpleChannelInboundHandler<MqttMessag
         }
     }
 
-    public MqttMessageUpContext buildMqttMessageUpContext(ChannelHandlerContext ctx) {
+    @Override
+    protected MqttMessageUpContext buildMqttMessageUpContext(ChannelHandlerContext ctx) {
         MqttMessageUpContext context = new MqttMessageUpContext();
         Channel channel = ctx.channel();
         context.setClientId(ChannelInfo.getClientId(channel));
